REF: sort_values, sort_index
